cddl: Do not unpack wrapped SUIT_Digests
The manifest authentication logic would like to both verify the digest,
which is encoded inside the SUIT_Digest structure, as well as verify the
signature of the whole SUIT_Digest structure, including the header
bytes.
To simplify the data transmission between the envelope decoding and
manifest validation, only the manifest digest stored as a byte string
-encoded SUIT_Digest structure will be used.

This change affects the component parameter (image-digest), so an
additional piece of decoding must be introduced there to keep the
platform API contract.

It does not affect the severable elements, because the SUIT_Digest
structure is used there directly, without wrapping it inside the byte
string.

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>