core: Move current component list to exec state
Due to the fact, that the list of active components relates to the
active manifest and the manifest may change during execution, the list
must change with the change of active manifest.
Since the execution is managing the list of active components, it was
easier to put the list of active components inside the execution state
instead of the manifest state.

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>